fix:(method) improved method for removing user

+++ src/config/db/databaseServise.ts
@@ -274,38 +274,44 @@ class DatabaseService {
   }
 
   async removeUser(id: number) {
-    const parser = await this.getParser(id);
-    const profile = await this.getProfile(id);
-    const dataParser = await this.getDataParser(id);
-    const kufarObjectIds = parser?.kufar?.kufarAds;
-    await KufarAd.deleteMany({ _id: { $in: kufarObjectIds } });
-    if (dataParser) await DataParser.deleteOne(dataParser?._id);
-    await Parser.deleteOne(parser?._id);
-    await Premium.deleteOne(profile?.premium?._id);
-    await Profile.deleteOne(profile?._id);
-    await User.deleteOne({ id });
-    await cache.removeCache(`user:${id}`);
-    const cacheLanguages = await cache.getCache('languages');
-    const cacheUsers = await cache.getCache('ids');
-    if (cacheLanguages) {
-      const parsedLanguages = JSON.parse(cacheLanguages);
-
-      if (parsedLanguages[id]) {
-        const { [id]: _, ...updatedLanguages } = parsedLanguages;
-        await cache.setCache('languages', updatedLanguages, this.TTL);
+    const user = await this.getUser(id);
+    if (user) {
+      const parser = await this.getParser(id);
+      const profile = await this.getProfile(id);
+      const dataParser = await this.getDataParser(id);
+      const kufarObjectIds = parser?.kufar?.kufarAds;
+      await KufarAd.deleteMany({ _id: { $in: kufarObjectIds } });
+      await User.deleteOne({ id });
+      if (profile) {
+        await Premium.deleteOne(profile?.premium?._id);
+        await Profile.deleteOne(profile?._id);
       }
-    }
-    if (cacheUsers) {
-      const userIds: number[] = JSON.parse(cacheUsers);
-      const filteredUsers = userIds.filter((userId) => userId !== id);
-      await cache.setCache('ids', filteredUsers, this.TTL);
-      await TelegramService.sendMessageToChat(
-        `${[
-          `🗑️ Пользователь с id: <b>${id}</b> был удален`,
-          `👥 Всего пользователей: <b>${(await User.find({})).length}</b>`,
-        ].join('\n')}`,
-      );
-    }
+      if (parser) await Parser.deleteOne(parser?._id);
+      if (dataParser) await DataParser.deleteOne(dataParser?._id);
+
+      await cache.removeCache(`user:${id}`);
+      const cacheLanguages = await cache.getCache('languages');
+      const cacheUsers = await cache.getCache('ids');
+      if (cacheLanguages) {
+        const parsedLanguages = JSON.parse(cacheLanguages);
+        if (parsedLanguages[id]) {
+          const { [id]: _, ...updatedLanguages } = parsedLanguages;
+          await cache.setCache('languages', updatedLanguages, this.TTL);
+        }
+      }
+      if (cacheUsers) {
+        const userIds: number[] = JSON.parse(cacheUsers);
+        const filteredUsers = userIds.filter((userId) => userId !== id);
+        await cache.setCache('ids', filteredUsers, this.TTL);
+        await TelegramService.sendMessageToChat(
+          `${[
+            `🗑️ Пользователь с id: <b>${id}</b> был удален`,
+            `👥 Всего пользователей: <b>${(await User.find({})).length}</b>`,
+          ].join('\n')}`,
+        );
+      }
+      console.error(`Заблокированный пользователь c ID:${id} был удален!`);
+    } else console.error(`Пользователь c ID:${id} не был найден.`);
   }
 
   async setUrlKufar(userId: number, url: string, urlId: number) {
